Success feedback on creating a post
