+++ client/src/components/post/CreateNewPost.js
@@ -1,11 +1,13 @@
 import React, { useState } from 'react';
 import { createPost } from '../../helper/post/index';
 import { isAuthenticated } from '../../helper/auth/index';
-
 import { toast } from 'react-toastify';
+import PostCard from './PostCard';
 
 export default function CreateNewPost() {
   const [content, setContent] = useState('');
+  const [isNewPostCreated, setIsNewPostCreated] = useState(false);
+  const [responseData, setResponseData] = useState();
   const { user, token } = isAuthenticated();
 
   const handleChange = (event) => {
@@ -17,41 +19,42 @@ export default function CreateNewPost() {
     //backend request fired
     createPost(user._id, token, { content }).then((data) => {
       if (data.error) {
-        //setError(true);
         toast.error('This is an error');
       } else {
-        //setError('');
-        //setSuccess(true);
-        //setcontent empty and button disabled
         setContent('');
+        setResponseData(data);
+        setIsNewPostCreated(true);
       }
     });
   };
 
   return (
-    <div className="postFormContainer">
-      <div className="userImageContainer">
-        <img src="/images/defaultProfilePic.jpg" alt="User's Profile Pic" />
+    <>
+      <div className="postFormContainer">
+        <div className="userImageContainer">
+          <img src="/images/defaultProfilePic.jpg" alt="User's Profile Pic" />
+        </div>
+        <div className="textareaContainer">
+          <form onSubmit={onSubmit}>
+            <textarea
+              value={content}
+              placeholder="what's happening?"
+              id="postTextArea"
+              onChange={handleChange}
+            ></textarea>
+            <div className="buttonContainer">
+              <button
+                type="submit"
+                disabled={content === '' ? true : false}
+                id="submitPostButton"
+              >
+                Post
+              </button>
+            </div>
+          </form>
+        </div>
       </div>
-      <div className="textareaContainer">
-        <form onSubmit={onSubmit}>
-          <textarea
-            value={content}
-            placeholder="what's happening?"
-            id="postTextArea"
-            onChange={handleChange}
-          ></textarea>
-          <div className="buttonContainer">
-            <button
-              type="submit"
-              disabled={content === '' ? true : false}
-              id="submitPostButton"
-            >
-              Post
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
+      {isNewPostCreated ? <PostCard post={responseData}></PostCard> : ''}
+    </>
   );
 }
